requirements.txt
streamlit
python-dotenv
openai
PyPDF2
pillow
pandas
fpdf
python-dateutil
requests